// src/app/interfaces/profesion.interface.ts
// ====================================================
// Interface Profesión permite la implementación rápida
// en otras interfaces o clases
// ====================================================

export interface ProfesionInterface {
    area?: string;
    titulo?: string;
    institucion?: string;
    fecha?: string;
    matProf?: string;
}